+++ server/storage-jdbc-clickhouse/src/main/java/org/bithon/server/storage/jdbc/clickhouse/ClickHouseStorageAutoConfiguration.java
@@ -19,7 +19,14 @@ package org.bithon.server.storage.jdbc.clickhouse;
 import com.alibaba.druid.pool.DruidDataSource;
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.Module;
-import org.bithon.server.storage.jdbc.clickhouse.web.DashboardStorage;
+import org.bithon.server.storage.jdbc.clickhouse.exception.ClickHouseExceptionTranslator;
+import org.bithon.server.storage.jdbc.clickhouse.storage.EventStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.MetadataStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.MetricStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.SchemaStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.SettingStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.TraceStorage;
+import org.bithon.server.storage.jdbc.clickhouse.storage.web.DashboardStorage;
 import org.jooq.impl.DSL;
 import org.jooq.impl.DefaultConfiguration;
 import org.jooq.impl.DefaultExecuteListenerProvider;
