+++ server/storage-jdbc-clickhouse/src/main/java/org/bithon/server/storage/jdbc/clickhouse/exception/ClickHouseExceptionTranslator.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.clickhouse;
+package org.bithon.server.storage.jdbc.clickhouse.exception;
 
 import org.jooq.ExecuteContext;
 import org.jooq.impl.DefaultExecuteListener;
@@ -25,7 +25,7 @@ import java.sql.SQLException;
  * @author frank.chen021@outlook.com
  * @date 2023/3/9 21:56
  */
-class ClickHouseExceptionTranslator extends DefaultExecuteListener {
+public class ClickHouseExceptionTranslator extends DefaultExecuteListener {
     @Override
     public void exception(ExecuteContext ctx) {
         SQLException exception = ctx.sqlException();
